moving to _cnab_util.py
