+++ partnercenter/azext_partnercenter/operations/marketplace_bundle/custom.py
@@ -11,6 +11,7 @@ import time
 import docker
 from azure.cli.core.util import sdk_no_wait
 from knack.util import CLIError
+from ._cnab_util import call_verify, call_buildbundle, _print_result, _get_container_id, _list_directory, _make_local_directory, _copy_to_local_directory, _login_to_acr, _get_running_container_id, _get_mount_path, _remove_container, _start_container, _stop_container
 
 # API Operations
 # pylint: disable=too-many-locals
@@ -21,48 +22,10 @@ def verify_bundle(manifest_file):
 
 def build_bundle(manifest_file):
     container_id = _get_container_id(manifest_file)
-    print("listing /root/.azure")
-    result = _list_directory(container_id, '/root/.azure')
-    _print_result(result)
-    result = _make_local_directory(container_id, '/cpalocal')
-    result = _list_directory(container_id, '/cpalocal')
-    result = _copy_to_local_directory(container_id, '/cpalocal')
-    print("listing cpalocal")
-    result = _list_directory(container_id, '/cpalocal')
-    _print_result(result)
-    print("listing cpalocal/cpaMount")
-    result = _list_directory(container_id, '/cpalocal/cpaMount')
-    _print_result(result)
     result = _login_to_acr(container_id)
-    _print_result(result)
     result = call_buildbundle(container_id, './cpalocal/cpaMount')
     _print_result(result)
     
-def call_verify(container_id, directory):
-    command = "docker exec " + container_id + " cpa verify --directory " + directory
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def call_buildbundle(container_id, directory):
-    #command = "docker exec " + container_id + " cpa buildbundle --directory " + directory
-    command = "docker exec -w /cpaMount " + container_id + " cpa buildbundle"
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _make_local_directory(container_id, directory):
-    command = "docker exec " + container_id + " mkdir " + directory
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _list_directory(container_id, directory):
-    command = "docker exec " + container_id + " ls -la " + directory
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _copy_to_local_directory(container_id, directory):
-    command = "docker exec " + container_id + " cp -R /cpaMount " + directory
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _login_to_acr(container_id):
-    command = "docker exec " + container_id + " az acr login -n bobjactranscontainers"
-    return subprocess.run([command],shell=True,capture_output=True)
-
 def update_bundle(cmd, instance, arg):
     # TODO: Implement partnercenter marketplace offer update
     return instance
@@ -76,48 +39,6 @@ def get_bundle(cmd):
 def list_bundle(cmd):
     raise CLIError('TODO: Implement `partnercenter marketplace offer show`')
 
-def _print_result(result):
-    output = result.stdout.decode("utf-8")
-    error = result.stderr.decode("utf-8")
-    print(output)
-
-def _get_container_id(manifest_file):
-    container_name = "cpacontainer"
-    container_lookup_name = "/" + container_name
-    container_id  = _get_running_container_id(container_lookup_name)
-
-    if container_id == "":
-        mount_path = _get_mount_path(manifest_file)
-        result = _start_container(container_name, mount_path)
-        container_id = _get_running_container_id(container_lookup_name)
-    
-    return container_id
-
-def _get_running_container_id(container_name):
-    client = docker.APIClient(base_url='unix://var/run/docker.sock')
-    ps = client.containers(False, True)
-    for item in ps:
-        if item['Names'][0] == container_name:
-            return item['Id']
-    return ""
-
-def _start_container(container_name, mount_path):
-    container_image = "bobjac/cnab:8.0"
-    #container_image = "mcr.microsoft.com/container-package-app:latest"
-    command = "docker run --name " + container_name + " -d -v /var/run/docker.sock:/var/run/docker.sock -v " + mount_path + ":/cpaMount" + " -v $HOME/.azure:/root/.azure "+ container_image + " sleep infinity"
-    print(command)
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _stop_container(container_name):
-    command = "docker stop " + container_name
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _remove_container(container_name):
-    command = "docker rm " + container_name
-    return subprocess.run([command],shell=True,capture_output=True)
-
-def _get_mount_path(manifest_file):
-    return os.path.dirname(manifest_file)
